@@ -1,39 +1,22 @@
 {
+  "browserslist": "> 0.5%, last 2 versions, not dead",
   "devDependencies": {
-    "babel-cli": "6.3.17",
-    "babel-core": "6.3.21",
-    "babel-loader": "6.2.0",
-    "babel-polyfill": "6.3.14",
-    "babel-preset-es2015": "6.3.13",
-    "babel-preset-react": "6.3.13",
-    "chai": "3.4.1",
-    "classnames": "2.2.1",
-    "css-loader": "0.23.1",
-    "cssnano": "3.4.0",
-    "extract-text-webpack-plugin": "0.9.1",
-    "node-sass": "3.8.0",
-    "normalize.css": "3.0.3",
-    "postcss-loader": "0.8.0",
-    "react-hot-loader": "1.3.0",
-    "redux-logger": "2.3.1",
-    "redux-thunk": "1.0.3",
-    "sass-loader": "3.1.2",
-    "style-loader": "0.13.0",
-    "webpack": "1.12.9",
-    "webpack-dev-server": "1.14.0"
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "react"
-    ]
+    "@parcel/transformer-sass": "2.12.0",
+    "@parcel/validator-typescript": "2.12.0",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "constants-browserify": "1.0.0",
+    "normalize.css": "8.0.1",
+    "parcel": "2.12.0",
+    "process": "0.11.10",
+    "typescript": "5.4.5"
   },
   "dependencies": {
-    "immutable": "3.7.6",
-    "react": "0.14.3",
-    "react-addons-pure-render-mixin": "0.14.3",
-    "react-dom": "0.14.3",
-    "react-redux": "4.0.3",
-    "redux": "3.0.5"
+    "classnames": "2.5.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-redux": "9.1.2",
+    "redux": "5.0.1",
+    "redux-thunk": "3.1.0"
   }
 }
