@@ -1,9 +1,9 @@
-import "lib/amplitude";
-import { LOCALSTORAGE, getItem, setItem } from "util/storage";
+import "../lib/amplitude";
+import { LOCALSTORAGE, getItem, setItem } from "./storage";
 
 const AMPLITUDE_API_KEY = "43c9261027eec7e7a89ceca5f5d4563d";
 const DO_NOT_TRACK =
-  __DEV__ ||
+  process.env.NODE_ENV !== "production" ||
   window.location.search.includes("dnt") ||
   getItem(LOCALSTORAGE.DNT);
 
@@ -13,7 +13,7 @@ if (DO_NOT_TRACK) {
   window.amplitude.init(AMPLITUDE_API_KEY, null, { includeReferrer: true });
 }
 
-export const track = (eventName, properties = {}) => {
+export const track = (eventName: string, properties = {}) => {
   if (DO_NOT_TRACK) {
     console.log(`Tracking ${eventName}:`, properties);
     return;
@@ -23,6 +23,6 @@ export const track = (eventName, properties = {}) => {
     Object.assign({}, properties, {
       screen_height: window.screen.height,
       screen_width: window.screen.width,
-    })
+    }),
   );
 };
